Changed the file and directory structure
